Update/locidex (#96)
* starting locidex database integrations

* updated added new files

* updated locidex db identification

* updated automated allele selection

* updated todos

* added test data for locidex databases, tests should still function:

* updated nextflow schema.json

* Removed tests for comparing updates for old and new locidex profiles

Perhaps the relevant fields should be sub set out

* removed dead test file

* updated test profile

* restrucutured file staging for locidex select

* started creation of real tests for locidex workflow

* updated locidex test data

* updated locidex workflow and tests

* updated docker container

* addressed PR issues, still more work to be done

* updated database selection functions to get an optimal match

* added skipped allele calling to test config

* updated test conditions

* updated test files

* updated locidex select paths

* removed trailing whitespace

* updated tests to include passing cases

* added locidex summary process

* updated fixe issues with integer sizing

* debugging summary test

* Added field for reportable alleles

* updated reportable loci test

* updated locidex summary tests

* updated missing allelels JSON type

* upated not database selected output

* updated IRIDANEXT config for locidex values

* updated locidex end-to-end tests

* updated locations of locidex test

* zipped and updated sample datasets

* upated locidex end to end tests

* fixed typos in change log

* updated locidex tests

* cleaned up datebase name parsing

* fixed code comment

* updated code comments

* made updates on PR comments

* normalized output for iridanext across allele scheme input options

* updated test cases

* updated locidex_select tests to match updated interface

* updated test to work in github actions

* removed sneaky todo

* updating issues identified and regression in kraken2 header parsing

* reverted zipped file handling

* fixed regression locidex select

* removed repeated variables

* updated staging of report input to locidex summarize

* updated file parsing in locidex summarize

Thank you Aaron!

* updated changelog

* extracted inline code in function

* bumped locidex container

* updated final todos

* updated options to locidex_select test to terminate processes on failure

+++ modules/local/locidex_search.nf
@@ -8,7 +8,6 @@ process LOCIDEX_SEARCH {
     label "process_medium"
     container "${workflow.containerEngine == 'singularity' || workflow.containerEngine == 'apptainer' ? task.ext.parameters.get('singularity') : task.ext.parameters.get('docker')}"
 
-
     input:
     tuple val(meta), path(fasta), path(db)
 
@@ -43,7 +42,7 @@ process LOCIDEX_SEARCH {
     --min_aa_match_cov ${params.locidex.min_aa_match_cov} \\
     --max_target_seqs ${params.locidex.max_target_seqs}
 
-    gzip -c seq_store.json > $output_json && rm seq_store.json
+    test -f seq_store.json && gzip -c seq_store.json > $output_json && rm seq_store.json
     test -f annotations.gbk && gzip -c annotations.gbk > $output_gbk && rm annotations.gbk
 
     cat <<-END_VERSIONS > versions.yml
